backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pydantic==1.10.14
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pandas==2.2.2
rapidfuzz==3.6.1
python-dotenv==1.0.1
unidecode==1.3.8
joblib==1.4.2
scikit-learn==1.4.2
sklearn-crfsuite==0.3.6
transformers==4.40.0
torch==2.2.2
openpyxl==3.1.2
redis==5.0.4